SAK-2887
Put the slide materials into an iFrame instead of an img tag.

This makes copyright materials work, pdfs work, etc.

Sometimes the browser gets a little confused when it gets redirected from a jpg
to a form - the workaround seems to be for viewers to preview the copyrighted
slides once before the show starts.

But it is much better.



git-svn-id: https://source.sakaiproject.org/svn/presentation/trunk@53140 66ffb92e-73f9-0310-93c1-f5514f145a0a